+++ templates/__common__/config/scripts/build.js
@@ -8,6 +8,7 @@ const colors = require('ansi-colors');
 const { logErrorMessage } = require('../utils');
 
 // tasks
+const lastBuildTime = require('../tasks/last-build-time');
 const api = require('../tasks/api');
 const clean = require('../tasks/clean');
 const copy = require('../tasks/copy');
@@ -21,6 +22,7 @@ const unusedCSS = require('../tasks/unused-css');
 
 async function build() {
   const runner = series([
+    lastBuildTime,
     clean,
     scripts,
     parallel([api, images, styles]),